requirements.txt
Flask==2.3.3
gunicorn==21.2.0
pymupdf==1.22.3
Pillow==10.3.0
